+++ api/usersGroup/delete.handler.php
@@ -14,6 +14,7 @@ if (!defined('IS_NOT_HACKED')) {
 }
 
 use \core\PHPLibrary\UserGroup as UserGroup;
+use \core\PHPLibrary\Users as Users;
 
 if ($system_core->client->is_logged(2)) {
   $client_user = $system_core->client->get_user(2);
@@ -27,14 +28,20 @@ if ($system_core->client->is_logged(2)) {
 
       if (UserGroup::exists_by_id($system_core, $user_group_id)) {
         $user_group = new UserGroup($system_core, $user_group_id);
+        $users = new Users($system_core);
 
-        if ($user_group_id > 4) {
-          $user_group_is_deleted = $user_group->delete();
-          if ($user_group_is_deleted) {
-            $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
-            $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+        if ($users->get_count_by_group_id($user_group_id) > 0) {
+          if ($user_group_id > 4) {
+            $user_group_is_deleted = $user_group->delete();
+            if ($user_group_is_deleted) {
+              $handler_message = (!isset($handler_message)) ? $system_core->locale->get_single_value_by_key('API_DELETE_DATA_SUCCESS') : $handler_message;
+              $handler_status_code = (!isset($handler_status_code)) ? 1 : $handler_status_code;
+            } else {
+              $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+              $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
+            }
           } else {
-            $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_ERROR_UNKNOWN')) : $handler_message;
+            $handler_message = (!isset($handler_message)) ? sprintf('API ERROR: %s', $system_core->locale->get_single_value_by_key('API_USERS_GROUP_ERROR_DELETION_EXISTS_USERS')) : $handler_message;
             $handler_status_code = (!isset($handler_status_code)) ? 0 : $handler_status_code;
           }
         } else {
